Make the widget system more React-like
