+++ ground-server/src/lib/definitions.ts
@@ -1,15 +1,20 @@
-import { type ReactNode } from "react";
+import { type ComponentType } from "react";
 import { type Layout } from "react-grid-layout";
 
 export type Data = {
   temp: number;
 };
 
+export type WidgetProps = {
+  mode: string;
+  data: Data;
+};
+
 export type TelemetryChannel = {
   label: string;
   requiresAuth: boolean;
   modes: string[];
-  render: (mode: string, data: Data) => ReactNode;
+  component: ComponentType<WidgetProps>;
 };
 
 export type Widget = {
