+++ ground-server/src/lib/telemetry-channels.ts
@@ -1,0 +1,12 @@
+import TemperatureWidget from "@/components/widgets/temperature-widget";
+
+import type { TelemetryChannel } from "@/lib/definitions";
+
+export const TELEMETRY_CHANNELS: TelemetryChannel[] = [
+  {
+    label: "Temperature",
+    requiresAuth: false,
+    modes: ["Value", "Chart"],
+    component: TemperatureWidget,
+  },
+];
